Se integra el tracking de consulta de productos para obtener el habito de consulta del cliente
