+++ package.json
@@ -37,6 +37,7 @@
     "react-toastify": "^10.0.4",
     "sitemap": "^8.0.0",
     "slick-carousel": "^1.8.1",
+    "uuid": "^10.0.0",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
@@ -45,6 +46,7 @@
     "@types/react": "^18.2.64",
     "@types/react-dom": "^18.2.21",
     "@types/react-slick": "^0.23.13",
+    "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
     "@vitejs/plugin-react": "^4.2.1",
