+++ src/components/LandingEcommerce/09Allies/Allies.tsx
@@ -1,38 +1,38 @@
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next';
 //ELEMENTOS DEL COMPONENTE
-import TDWire from '../../assets/Landing/06Allies/TDWire.png';
-import Southwire from '../../assets/Landing/06Allies/Southwire.png';
-import Prysmian from '../../assets/Landing/06Allies/Prysmian.png';
-import Teldor from '../../assets/Landing/06Allies/Teldor.png';
-import Aristoncavi from '../../assets/Landing/06Allies/Aristoncavi.png';
-import Emcocables from '../../assets/Landing/06Allies/Empocables.png';
-import ZTT from '../../assets/Landing/06Allies/ZTT.png';
-import Hengtong from '../../assets/Landing/06Allies/Hengtong.png';
-import FutureFibresTechnologies from '../../assets/Landing/06Allies/FutureFibresTechnologies.png';
-import Mecano from '../../assets/Landing/06Allies/Mecano.png';
-import Tenaris from '../../assets/Landing/06Allies/Tenaris.png';
-import Tecna from '../../assets/Landing/06Allies/Tecna.png';
-import Roxtec from '../../assets/Landing/06Allies/Roxtec.png';
-import MCS from '../../assets/Landing/06Allies/MCS.png';
-import TEConnectivity from '../../assets/Landing/06Allies/TEConnectivity.png';
-import BornasTerminales from '../../assets/Landing/06Allies/BornasTerminales.png';
-import SoldexEl from '../../assets/Landing/06Allies/SoldexEl.png';
-import Dialight from '../../assets/Landing/06Allies/Dialight.png';
-import Sylvania from '../../assets/Landing/06Allies/Sylvania.png';
-import AEL from '../../assets/Landing/06Allies/AEL.png';
-import Lithonia from '../../assets/Landing/06Allies/Lithonia.png';
-import Rockwell from '../../assets/Landing/06Allies/Rockwell.png';
-import Siemens from '../../assets/Landing/06Allies/Siemens.png';
-import ABB from '../../assets/Landing/06Allies/ABB.png';
-import Weidmuller from '../../assets/Landing/06Allies/Weidmuller.png';
-import AcuityControls from '../../assets/Landing/06Allies/AcuityControls.png';
-import Omnio from '../../assets/Landing/06Allies/Omnio.png';
-import Rittal from '../../assets/Landing/06Allies/Rittal.png';
-import Dyminel from '../../assets/Landing/06Allies/Dyminel.jpg';
-import ColdJet from '../../assets/Landing/06Allies/ColdJet.png';
-import TresM from '../../assets/Landing/06Allies/3M.png';
-import Adler from '../../assets/Landing/06Allies/Adler.png';
+import TDWire from '../../../assets/Landing/06Allies/TDWire.png';
+import Southwire from '../../../assets/Landing/06Allies/Southwire.png';
+import Prysmian from '../../../assets/Landing/06Allies/Prysmian.png';
+import Teldor from '../../../assets/Landing/06Allies/Teldor.png';
+import Aristoncavi from '../../../assets/Landing/06Allies/Aristoncavi.png';
+import Emcocables from '../../../assets/Landing/06Allies/Empocables.png';
+import ZTT from '../../../assets/Landing/06Allies/ZTT.png';
+import Hengtong from '../../../assets/Landing/06Allies/Hengtong.png';
+import FutureFibresTechnologies from '../../../assets/Landing/06Allies/FutureFibresTechnologies.png';
+import Mecano from '../../../assets/Landing/06Allies/Mecano.png';
+import Tenaris from '../../../assets/Landing/06Allies/Tenaris.png';
+import Tecna from '../../../assets/Landing/06Allies/Tecna.png';
+import Roxtec from '../../../assets/Landing/06Allies/Roxtec.png';
+import MCS from '../../../assets/Landing/06Allies/MCS.png';
+import TEConnectivity from '../../../assets/Landing/06Allies/TEConnectivity.png';
+import BornasTerminales from '../../../assets/Landing/06Allies/BornasTerminales.png';
+import SoldexEl from '../../../assets/Landing/06Allies/SoldexEl.png';
+import Dialight from '../../../assets/Landing/06Allies/Dialight.png';
+import Sylvania from '../../../assets/Landing/06Allies/Sylvania.png';
+import AEL from '../../../assets/Landing/06Allies/AEL.png';
+import Lithonia from '../../../assets/Landing/06Allies/Lithonia.png';
+import Rockwell from '../../../assets/Landing/06Allies/Rockwell.png';
+import Siemens from '../../../assets/Landing/06Allies/Siemens.png';
+import ABB from '../../../assets/Landing/06Allies/ABB.png';
+import Weidmuller from '../../../assets/Landing/06Allies/Weidmuller.png';
+import AcuityControls from '../../../assets/Landing/06Allies/AcuityControls.png';
+import Omnio from '../../../assets/Landing/06Allies/Omnio.png';
+import Rittal from '../../../assets/Landing/06Allies/Rittal.png';
+import Dyminel from '../../../assets/Landing/06Allies/Dyminel.jpg';
+import ColdJet from '../../../assets/Landing/06Allies/ColdJet.png';
+import TresM from '../../../assets/Landing/06Allies/3M.png';
+import Adler from '../../../assets/Landing/06Allies/Adler.png';
 import styles from './styles.module.css';
 
 function Allies() {
