+++ src/components/LandingEcommerce/09Allies/styles.module.css
@@ -1,5 +1,4 @@
 .container {
-    margin: 80px auto;
     max-width: 1200px;
 }
 .main__Title {
