+++ src/pages/LandingEcommerce/03BestSellingProducts/BestSellingProductsPage.tsx
@@ -3,13 +3,13 @@ import Footer from '../../../components/LandingEcommerce/Footer/Footer';
 import ConstructionWorker from '../../../assets/Error404/ConstructionWorker.svg';
 import styles from './styles.module.css';
 
-function TopSellersPage() {
+function BestSellingProductsPage() {
     
     return (
         <div>
             <NavBar />
             <div className={`${styles.container} d-flex flex-column align-items-center justify-content-center`}>
-                <h1 className={`${styles.main__Title} m-0 text-center`}>TopSellersPage</h1>
+                <h1 className={`${styles.main__Title} m-0 text-center`}>BestSellingProductsPage</h1>
                 <div>
                     <img src={ConstructionWorker} alt="Página en construcción" className={styles.image} loading="lazy" />
                     <h2 className='text-center'>Página en construcción</h2>
@@ -20,4 +20,4 @@ function TopSellersPage() {
     );
 }
 
-export default TopSellersPage;
+export default BestSellingProductsPage;
