+++ src/pages/LandingEcommerce/05ProductsOnOffer/ProductsOnOfferPage.tsx
@@ -3,13 +3,13 @@ import Footer from '../../../components/LandingEcommerce/Footer/Footer';
 import ConstructionWorker from '../../../assets/Error404/ConstructionWorker.svg';
 import styles from './styles.module.css';
 
-function OffersPage() {
+function ProductsOnOfferPage() {
 
     return (
         <div>
             <NavBar />
             <div className={`${styles.container} d-flex flex-column align-items-center justify-content-center`}>
-                <h1 className={`${styles.main__Title} m-0 text-center`}>OffersPage</h1>
+                <h1 className={`${styles.main__Title} m-0 text-center`}>ProductsOnOfferPage</h1>
                 <div>
                     <img src={ConstructionWorker} alt="Página en construcción" className={styles.image} loading="lazy" />
                     <h2 className='text-center'>Página en construcción</h2>
@@ -20,4 +20,4 @@ function OffersPage() {
     );
 }
 
-export default OffersPage;
+export default ProductsOnOfferPage;
