+++ src/pages/LandingEcommerce/DetailProduct/DetailProductPage.tsx
@@ -1,12 +1,14 @@
 /* eslint-disable react-hooks/exhaustive-deps, @typescript-eslint/no-explicit-any */
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { useParams } from 'react-router-dom';
+import jsCookie from "js-cookie";
 import { pdf } from '@react-pdf/renderer';
 import { saveAs } from 'file-saver';
 // REDUX
 import { useDispatch, useSelector } from 'react-redux';
 import type { RootState, AppDispatch } from '../../../redux/store';
 import { getProductById } from '../../../redux/LandingEcommerce/productSlice/actions';
+import { postTrackProductView } from '../../../redux/LandingEcommerce/productSlice/actions';
 // ELEMENTOS DEL COMPONENTE
 import NavBar from '../../../components/LandingEcommerce/01NavBar/NavBar';
 import Footer from '../../../components/LandingEcommerce/Footer/Footer';
@@ -15,20 +17,22 @@ import ModalShopingCard from '../../../components/LandingEcommerce/DetailProduct
 // import Accesories from '../../../components/LandingEcommerce/Accesories/Accesories';
 import Loading from '../../../components/ComponentLoading/Loading';
 import { IProduct } from '../../../types/product.types';
+import { IInspiredByLastSaw } from '../../../types/InspiredByLastSaw.types';
 import { ICartProduct } from '../../../types/cartProduct';
 import { IFavorites } from '../../../types/favorites.types';
 import { formatNumber } from '../../../helpers/FormatNumber/FormatNumber';
 import SliderZoom from '../../../helpers/SliderZoom/SliderZoom';
+import TechnicaData from '../../../components/LandingEcommerce/DetailProduct/TechnicaData/TechnicaData';
 import { AiFillFilePdf } from "react-icons/ai";
 import { BsDownload } from "react-icons/bs";
 import { GrStar } from 'react-icons/gr';
 import { VscStarHalf } from "react-icons/vsc";
 import { HiMiniPlus, HiMiniMinus } from "react-icons/hi2";
 import { IoMdHeart , IoMdHeartEmpty } from "react-icons/io";
-import { PiCaretUp, PiCaretDown } from "react-icons/pi";
 import styles from './styles.module.css';
 
 function DetailProductPage() {
+    const token = jsCookie.get('token') || '';
     const { idProduct } = useParams<{ idProduct: string }>();
     const dispatch: AppDispatch = useDispatch();
     const [downloadPdf, setDownloadPdf] = React.useState(false);
@@ -38,9 +42,8 @@ function DetailProductPage() {
     const user = useSelector((state: RootState) => state.user.user);
     const errorProduct = useSelector((state: RootState) => state.products.errorProduct);
     
-    const idUser = user?.id;                                                // Id del cliente>(null);
-    const [showTechnicalData, setShowTechnicalData] = useState(false);      //Expande la ficha técnica
-    const [count, setCount] = useState(1);                                  // Cantidad seleccionada de cada producto
+    const idUser = user?.id;
+    const [count, setCount] = useState(1);
     const [showModal, setShowModal] = useState(false);
     
     const [productsInLocalStorage, setProductsInLocalStorage] = useState<ICartProduct[]>();
@@ -50,12 +53,25 @@ function DetailProductPage() {
         client: idUser,
     });
 
+    //GUARDA LA TRAZABILIDAD DE CONSULTA DEL CLIENTE
+    const isTracked = useRef(false);
     useEffect(() => {
-        if (idProduct) {
+        if (idProduct && !isTracked.current) {
             dispatch(getProductById(idProduct));
+            const sessionId = sessionStorage.getItem('sessionId');
+            if (sessionId) {
+                const formData: IInspiredByLastSaw = {
+                    sessionId,
+                    productId: idProduct,
+                    client: token || '',
+                };
+                dispatch(postTrackProductView(formData));
+                isTracked.current = true;
+            }
         }
     }, [idProduct]);
-
+    
+    
     const handleIncrement = () => {
         setCount(prevCount => prevCount + 1);
     };
@@ -67,8 +83,7 @@ function DetailProductPage() {
     useEffect(() => {
         if (productsState && !Array.isArray(productsState)) {
             const product = productsState as IProduct;
-            const sellingPrice = product.promotionalPrice ?? product.sellingPriceFinalUser ?? 0; // Asigna 0 si ambos son undefined
-    
+            const sellingPrice = product.promotionalPrice ?? product.sellingPriceFinalUser ?? 0;
             const mainProduct: ICartProduct = {
                 productId: product._id,
                 sap: product.sap,
@@ -83,9 +98,9 @@ function DetailProductPage() {
                 description: product.description,
                 detailUrl: `/details/${idProduct}`,
                 quantity: count,
-                iva: product.iva, // Asegurado de ser un número
-                sellingPrice: sellingPrice, // Asegurado de ser un número
-                subtotal: count * sellingPrice, // Asegurado de no ser undefined
+                iva: product.iva,
+                sellingPrice: sellingPrice,
+                subtotal: count * sellingPrice,
             };
             
             setSelectedProducts(prevState => ({
@@ -97,16 +112,12 @@ function DetailProductPage() {
     }, [productsState, count]);
 
     const handleAddToCart = () => {
-        // Obtener carrito desde el localStorage
-        const storedCart = localStorage.getItem('order');
+        const storedCart = localStorage.getItem('order');               // Obtener carrito desde el localStorage
         const cart = storedCart ? JSON.parse(storedCart) : { products: [], total: 0, client: idUser };
-        
-        // Agregar el producto actual al carrito
-        const product = selectedProducts.products[0];
+        const product = selectedProducts.products[0];                   // Agregar el producto actual al carrito
         const existingProductIndex = cart.products.findIndex((p: { productId: string | undefined; }) => p.productId === product.productId);
         
-        if (existingProductIndex !== -1) {
-            // Si el producto ya existe, incrementar la cantidad y actualizar subtotal
+        if (existingProductIndex !== -1) {                              // Si el producto ya existe, incrementar la cantidad y actualizar subtotal
             cart.products[existingProductIndex].quantity += product.quantity;
             cart.products[existingProductIndex].subtotal += product.subtotal;
         } else {
@@ -131,7 +142,7 @@ function DetailProductPage() {
     }, []);
 
     const handleAddFavorite = () => {
-        if (!product) return; // Verificar que el producto esté cargado
+        if (!product) return;
         const newFavorite: IFavorites = {
             productId: product._id,
             sap: product.sap,
@@ -156,7 +167,6 @@ function DetailProductPage() {
         localStorage.setItem('favoriteProducts', JSON.stringify([...favoriteProducts, newFavorite]));       // Guardar en localStorage
     };
 
-    // Función para cambiar el estado y mostrar/ocultar el modal
     const toggleModal = () => {
         setShowModal(true);
     };
@@ -175,7 +185,7 @@ function DetailProductPage() {
                 );
                 const blob = await pdf(<MyDocument />).toBlob();
                 saveAs(blob, 'ficha_tecnica.pdf');
-                setDownloadPdf(false); // Resetear el estado después de la descarga
+                setDownloadPdf(false);
             };
             generatePdfDocument();
         }
@@ -190,15 +200,6 @@ function DetailProductPage() {
         productsState.secondaryImage,
         ...(productsState.otherImages ?? [])
     ];
-
-    // Expansión de la Ficha técnica
-    function handleShowBanks() {
-        setShowTechnicalData(true);
-    }
-
-    function handleHideBanks() {
-        setShowTechnicalData(false);
-    }
 
     const handleDownload = () => {
         setDownloadPdf(true);
@@ -217,7 +218,7 @@ function DetailProductPage() {
                             <GrStar className={`${styles.star__Score}`} /><GrStar className={`${styles.star__Score}`} /><GrStar className={`${styles.star__Score}`} /><GrStar className={`${styles.star__Score}`} /><VscStarHalf className={`${styles.star__Score}`} /> 4.7 (50)
                         </div>
                         <p className={`${styles.price__Product} m-0`}>$ {formatNumber(product.sellingPriceFinalUser)} und</p>
-                        <p className={`${styles.stock__Product} m-0`}>{product.inventory} unidades en stock</p>
+                        <p className={`${styles.stock__Product} m-0`}>{product.inventory} unidades en stockK</p>
 
                         <div className={`${styles.container__Updated_Cart} d-flex align-items-center justify-content-between`}>
                             <div className={`${styles.container__Count_Favorites} d-flex align-items-center justify-content-between`}>
@@ -291,227 +292,11 @@ function DetailProductPage() {
                     </div>
                 </div>
 
-                {/* <Accesories /> */}
+                {/* COMPONENTE DE <Accesories /> */}
 
-                <div className={`${styles.container__Technical_Data} mt-4`}>
-                    <div className={`${styles.technical__Data} pt-1 pb-1 px-0 d-flex align-items-center justify-content-between`}>
-                        <h3 className={styles.tittle__Technical_Data}>Características técnicas</h3>
-                        {showTechnicalData === false && (
-                            <PiCaretDown className={`${styles.icon__Show} `} onClick={handleShowBanks} />
-                        )}
-                        {showTechnicalData === true && (
-                            <PiCaretUp className={`${styles.icon__Show} `} onClick={handleHideBanks} />
-                        )}
-                    </div>
-
-                    {showTechnicalData && (
-                        <div className="pt-2">
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Información general</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Código SAP</div>
-                                    <div className={styles.value__Characteristics}>{product.sap}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Clase</div>
-                                    <div className={styles.value__Characteristics}>{product.class}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Categoría</div>
-                                    <div className={styles.value__Characteristics}>{product.category}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Tipo</div>
-                                    <div className={styles.value__Characteristics}>{product.type}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Grupo</div>
-                                    <div className={styles.value__Characteristics}>{product.group}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Unidad de medida</div>
-                                    <div className={styles.value__Characteristics}>{product.unitMeasure}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Fabricante</div>
-                                    <div className={styles.value__Characteristics}>{product.manufacturer}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Familia</div>
-                                    <div className={styles.value__Characteristics}>{product.family}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Serie</div>
-                                    <div className={styles.value__Characteristics}>{product.series}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Referencia</div>
-                                    <div className={styles.value__Characteristics}>{product.reference}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Activo</div>
-                                    <div className={styles.value__Characteristics}>{product.active}</div>
-                                </div>
-                            </div>
-                            
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Características constructivas</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Código SAP</div>
-                                    <div className={styles.value__Characteristics}>{product.sap}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Características eléctricas</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Tipo de tensión</div>
-                                    <div className={styles.value__Characteristics}>{product.tensionType}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Tensión mínima</div>
-                                    <div className={styles.value__Characteristics}>{product.minimumTension}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Tensión máxima</div>
-                                    <div className={styles.value__Characteristics}>{product.maximumTension}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Corriente de entrada</div>
-                                    <div className={styles.value__Characteristics}>{product.inputCurrent}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Corriente de salida</div>
-                                    <div className={styles.value__Characteristics}>{product.outputCurrent}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Potencia</div>
-                                    <div className={styles.value__Characteristics}>{product.power}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Voltaje de potencia</div>
-                                    <div className={styles.value__Characteristics}>{product.powerVoltage}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Caballaje</div>
-                                    <div className={styles.value__Characteristics}>{product.horsepower}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Voltaje de caballaje</div>
-                                    <div className={styles.value__Characteristics}>{product.horsepowerVoltage}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Frecuencia</div>
-                                    <div className={styles.value__Characteristics}>{product.frequency}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Características mecánicas</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Polos</div>
-                                    <div className={styles.value__Characteristics}>{product.poles}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Tamaño</div>
-                                    <div className={styles.value__Characteristics}>{product.size}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Alto</div>
-                                    <div className={styles.value__Characteristics}>{product.high}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Ancho</div>
-                                    <div className={styles.value__Characteristics}>{product.width}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Profundo</div>
-                                    <div className={styles.value__Characteristics}>{product.deep}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Peso</div>
-                                    <div className={styles.value__Characteristics}>{product.weight}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Montaje</div>
-                                    <div className={styles.value__Characteristics}>{product.mounting}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Conexión</div>
-                                    <div className={styles.value__Characteristics}>{product.connection}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Condiciones ambientales</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Eficiencia</div>
-                                    <div className={styles.value__Characteristics}>{product.efficiency}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>IP</div>
-                                    <div className={styles.value__Characteristics}>{product.ip}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Dimensiones</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Polos</div>
-                                    <div className={styles.value__Characteristics}>{product.poles}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Normas y estándares</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Norma</div>
-                                    <div className={styles.value__Characteristics}>{product.standard}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Protocolo</div>
-                                    <div className={styles.value__Characteristics}>{product.protocol}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Documentación</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Aplicaciones</div>
-                                    <div className={styles.value__Characteristics}>{product.applications}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Página</div>
-                                    <div className={styles.value__Characteristics}>{product.page}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Web</div>
-                                    <div className={styles.value__Characteristics}>{product.web}</div>
-                                </div>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Datasheet</div>
-                                    <div className={styles.value__Characteristics}>{product.datasheet}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Accesorios</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Aplicaciones</div>
-                                    <div className={styles.value__Characteristics}>{product.applications}</div>
-                                </div>
-                            </div>
-
-                            <div className={`${styles.container__Characteristics} mb-4`}>
-                                <h5 className={styles.title__Characteristics}>Homólogos</h5>
-                                <div className={`${styles.container__Propertie} d-flex`}>
-                                    <div className={styles.title__Propertie}>Aplicaciones</div>
-                                    <div className={styles.value__Characteristics}>{product.applications}</div>
-                                </div>
-                            </div>
-                        </div>
-                    )}
-                </div>
+                <TechnicaData
+                    product={product}
+                />
             </div>
             <Footer />
         </div>
