+++ src/types/InspiredByLastSaw.types.ts
@@ -1,0 +1,5 @@
+export interface IInspiredByLastSaw {
+    sessionId: string;
+    productId: string;
+    client?: string;
+}
